Fixing infinite loop
One of the limitations of this gem is that if you define a method with
hooks in a parent class and you use `super` in the same decorated method
in children you enter in an infinite loop.

This PR mitigates this situation by making sure that only the methods
included in hooks are decorated.