Use prepend instead of alias_method_chain like approach
Current implementation is using an old approach of renaming
decorated methods. This approach has some drawback, like the inhability
to call super, confusing method names, and other issues explanined here
by Justin Weiss
https://www.justinweiss.com/articles/rails-5-module-number-prepend-and-the-end-of-alias-method-chain/

The solution was proposed in Rails 5: using Ruby's prepend module to
define the methods that need to be decorated.

This commit creates a new `ClassName::CaptainHookDecorator` module in
each class where `hook` is called, prepends it to the class, and defines
the relevant decorator methods in that module